+++ ui/src/views/Game/GameView.tsx
@@ -1,25 +1,17 @@
 import React from "react";
-import Navbar from "../../component/Navbar";
+import Navbar from "../../components/Navbar";
 import { Box, Typography } from "@mui/material";
 import { translationKeys } from "./constants";
 import { useTranslation } from "react-i18next";
+import { makeStyles } from "tss-react/mui";
 
 export default function ProfileView(): React.ReactElement {
   const { t } = useTranslation();
+  const { classes } = useStyles();
   return (
     <>
       <Navbar />
-      <Box
-        sx={{
-          marginTop: "4rem",
-          border: "1px",
-          width: "100%",
-          height: "calc(100vh - 4rem)",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-      >
+      <Box className={classes.wrapperCard}>
         <Box
           sx={{
             background: "#fff1e1",
@@ -67,3 +59,16 @@ export default function ProfileView(): React.ReactElement {
     </>
   );
 }
+
+const useStyles = makeStyles()(() => ({
+  wrapperCard: {
+    marginTop: "4rem",
+    border: "1px",
+    width: "100%",
+    height: "calc(100vh - 4rem)",
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  inputName: {},
+}));
